backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.3.0
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
drf-yasg==1.21.7



